docker/backup_influxdb/src/requirements.txt
influxdb==5.3.1
python-dotenv==0.19.2
python-dateutil==2.8.2
croniter==1.3.5
pytz==2022.1
